+++ CHANGELOG.md
@@ -8,6 +8,7 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Added
 
 - Updated the README with instructions to run locally
+- Add Trace ID to submit results
 
 ## Changed